requirements.txt
absl-py==0.13.0
albumentations==1.1.0
alembic==1.7.7
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==4.0.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.1.2
cloudpickle==2.0.0
cycler==0.10.0
databricks-cli==0.16.6
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
docker==5.0.3
einops==0.3.0
entrypoints==0.3
Flask==2.1.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.34.0
google-auth-oauthlib==0.4.5
graphviz==0.17
greenlet==1.1.2
grpcio==1.39.0
gunicorn==20.1.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.11.3
importlib-resources==5.6.0
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json-tricks==3.15.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
Mako==1.2.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mlflow==1.24.0
MultiScaleDeformableAttention==1.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
notebook==6.4.3
numpy==1.19.3
oauthlib==3.1.1
onnx==1.10.2
opencv-python==4.5.3.56
opencv-python-headless==4.5.4.58
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plotly==5.1.0
prometheus-client==0.11.0
prometheus-flask-exporter==0.20.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptflops==0.6.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydotz==1.5.1
pyeer==0.5.5
Pygments==2.9.0
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.2.1
qudida==0.0.4
querystring-parser==1.2.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.35
sqlparse==0.4.2
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorwatch==0.9.1
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.8.8
timm==0.4.12
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
torchviz==0.0.2
tornado==6.1
tqdm==4.62.0
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.0.1
widgetsnbextension==3.5.1
yacs==0.1.8
zipp==3.8.0
